Support continuous recompiling of CSS with a looping compiler
Pass the new option --loop-after=X to the twbs:compileless task
to have it pause and then loop after the specified number of
seconds. Makes it quicker to iterate when working on the less files
without having to try and watch all the possible source files for
changes (which isn't possible over a virtualbox share anyway).